PlayerMaxBuffOverrides -> Tweaks/MaxBuffsOverride
